pyproject.toml
[project]
name = "cks_picks_cfb"
version = "0.1.0"
description = "An end-to-end college football betting model that predicts point spreads and over/unders."
requires-python = ">=3.12,<3.13"
authors = [{ name = "Connor Kitchings", email = "connor.kitchings@gmail.com" }]
dependencies = [
    "pandas>=2.0",
    "scikit-learn>=1.4",
    "streamlit>=1.30",
    "plotly>=5.18.0",
    "requests>=2.30",
    "python-dotenv>=1.0",
    # "prefect>=2.19.0",
    "shap>=0.50.0",
    "cfbd>=5.9.1", # CollegeFootballData API access
    "pyarrow>=15.0", # Parquet IO
    "jinja2>=3.1", # Email template rendering
    "mlflow>=2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "optuna<3.0.0,>=2.10.0",
    "lightgbm>=4.5.0",
    "catboost>=1.2.7",
    "geopy>=2.4.1",
    "iniconfig>=2.1.0",
    "pluggy>=1.6.0",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
    "xgboost>=1.5.0",
    "pymc>=5.0.0",
    "tabulate>=0.9.0",
    "typer>=0.20.0",
    "rich>=13.9.4",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "boto3>=1.35.0", # AWS/R2 cloud storage
]

[project.optional-dependencies]
dev = [
    "rich>=13.0",
    "ruff>=0.12.7",
    "ipykernel==6.29.5",
    "pytest>=8.4.1",
    "pandas-stubs==2.2.0.240218",
    "types-requests==2.31.0.20240218",
    "mkdocs-material==9.5.7",
    "mkdocs==1.5.3",
    "mike==2.0.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.8.0",
    "typer>=0.9.0",
]

[tool.ruff]
line-length = 88
extend-exclude = ["*.ipynb", "**/*.ipynb"]
force-exclude = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]



[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mkdocs-material>=9.5.7",
    "mkdocstrings[python]>=0.28.3",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "typer>=0.9.0",
    "bandit>=1.9.3",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
norecursedirs = [
  "scripts/archive",
  "legacy"
]
